Refactor `Body2D`
